Introduce a helper script for preparing RCs.
This is a bit piecemeal at the moment, but automates preparation of RC
artifacts created with `./build-artifact` for upload to bintray.

Basic docs are provided for preparing an RC, but more glue scripting is
still needed to make the process as smooth and automated as the main
aurora release.

Reviewed at https://reviews.apache.org/r/44530/