Remove some dependencies linked to docker usage
An operator should be able to build Aurora on his platform without
relying on Docker.

Even if this is an interesting way of doing it, it sometimes introduce missed
(build-)dependencies (in this case openssl) and context-specific behaviour
(here, the way pants.ini was added to the build context)

Testing Done:
Able to build on CentOS 6 using mock.
Able to build on Debian. Limitation = need the patched version of Gradle.
Still able to build using the current Docker build system

Bugs closed: AURORA-1568

Reviewed at https://reviews.apache.org/r/45298/